+++ Oipie/Favourites/Infrastructure/FavouritesViewController.swift
@@ -1,0 +1,29 @@
+//
+//  FavouritesViewController.swift
+//  Oipie
+//
+//  Created by Daniel Ramos on 30/8/22.
+//
+
+import UIKit
+
+class FavouritesViewController: UIViewController {
+    let label: UILabel = {
+        let title = UILabel()
+        title.text = "Favourites"
+        title.textAlignment = .center
+        return title
+    }()
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        view.backgroundColor = .systemPink
+        view.addSubview(label)
+    }
+
+    override func viewDidLayoutSubviews() {
+        label.center = view.center
+        label.frame = view.bounds
+    }
+}
