+++ Oipie/Home/Infrastructure/UI/HomeViewController.swift
@@ -1,0 +1,29 @@
+//
+//  HomeViewController.swift
+//  Oipie
+//
+//  Created by Daniel Ramos on 30/8/22.
+//
+
+import UIKit
+
+class HomeViewController: UIViewController {
+    let label: UILabel = {
+        let title = UILabel()
+        title.text = "Home"
+        title.textAlignment = .center
+        return title
+    }()
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        view.backgroundColor = .systemBlue
+        view.addSubview(label)
+    }
+
+    override func viewDidLayoutSubviews() {
+        label.center = view.center
+        label.frame = view.bounds
+    }
+}
