+++ Oipie/MyRecepies/Infrastructure/UI/MyRecepiesViewController.swift
@@ -1,0 +1,29 @@
+//
+//  MyRecepiesViewController.swift
+//  Oipie
+//
+//  Created by Daniel Ramos on 30/8/22.
+//
+
+import UIKit
+
+class MyRecepiesViewController: UIViewController {
+    let label: UILabel = {
+        let title = UILabel()
+        title.text = "Done Recepies"
+        title.textAlignment = .center
+        return title
+    }()
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        view.backgroundColor = .systemGreen
+        view.addSubview(label)
+    }
+
+    override func viewDidLayoutSubviews() {
+        label.center = view.center
+        label.frame = view.bounds
+    }
+}
