+++ Oipie/Profile/Infrastructure/UI/ProfileViewController.swift
@@ -1,0 +1,29 @@
+//
+//  ProfileViewController.swift
+//  Oipie
+//
+//  Created by Daniel Ramos on 30/8/22.
+//
+
+import UIKit
+
+class ProfileViewController: UIViewController {
+    let label: UILabel = {
+        let title = UILabel()
+        title.text = "Profile"
+        title.textAlignment = .center
+        return title
+    }()
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        view.backgroundColor = .systemBrown
+        view.addSubview(label)
+    }
+
+    override func viewDidLayoutSubviews() {
+        label.center = view.center
+        label.frame = view.bounds
+    }
+}
