+++ Oipie/Search/Infrastructure/SearchViewController.swift
@@ -1,0 +1,29 @@
+//
+//  SearchViewController.swift
+//  Oipie
+//
+//  Created by Daniel Ramos on 30/8/22.
+//
+
+import UIKit
+
+class SearchViewController: UIViewController {
+    let label: UILabel = {
+        let title = UILabel()
+        title.text = "Search"
+        title.textAlignment = .center
+        return title
+    }()
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        view.backgroundColor = .systemRed
+        view.addSubview(label)
+    }
+
+    override func viewDidLayoutSubviews() {
+        label.center = view.center
+        label.frame = view.bounds
+    }
+}
